Move Universal Tags structs to separate module (#820)
