+++ nexus-webapi/src/routes/v0/resource/mod.rs
@@ -8,9 +8,10 @@ use axum::extract::{Path, Query};
 use axum::routing::get;
 use axum::{Json, Router};
 use nexus_common::models::resource::tag::TagResource;
-use nexus_common::models::resource::{normalize_uri, resource_id, ResourceDetails};
+use nexus_common::models::resource::ResourceDetails;
 use nexus_common::models::tag::traits::{TagCollection, TaggersCollection};
 use nexus_common::models::tag::TagDetails;
+use nexus_common::universal_tag::normalize::{normalize_uri, resource_id};
 use serde::{Deserialize, Serialize};
 use tracing::debug;
 use utoipa::{OpenApi, ToSchema};
@@ -129,8 +130,7 @@ pub async fn resource_by_uri_handler(
 
     debug!("GET {RESOURCE_BY_URI_ROUTE} uri:{}", query.uri);
 
-    let (normalized, _scheme) =
-        normalize_uri(&query.uri).map_err(|e| Error::InvalidInput { message: e })?;
+    let (normalized, _scheme) = normalize_uri(&query.uri).map_err(Error::invalid_input)?;
     let res_id = resource_id(&normalized);
 
     let tags = TagResource::get_by_id(
